all: support post editing
