+++ go.mod
@@ -3,6 +3,7 @@ module changkun.de/x/redir
 go 1.16
 
 require (
+	github.com/google/uuid v1.2.0
 	go.mongodb.org/mongo-driver v1.4.6
 	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776
 )
