+++ short.go
@@ -146,6 +146,15 @@ func shortCmd(ctx context.Context, operate op, r *models.Redirect) (err error) {
 		}
 	}()
 
+	err = shortEdit(ctx, s, operate, r.Alias, r)
+	return
+}
+
+// shortEdit edits the datastore for a given alias in a given operation.
+// if the operation is create, then the alias is not necessary.
+// if the operation is update/fetch/delete, then the alias is used to
+// match the existing aliases, meaning that alias can be changed.
+func shortEdit(ctx context.Context, s *db.Store, operate op, a string, r *models.Redirect) (err error) {
 	switch operate {
 	case opCreate:
 		err = s.StoreAlias(ctx, r)
@@ -167,8 +176,14 @@ func shortCmd(ctx context.Context, operate op, r *models.Redirect) (err error) {
 
 		// fetch the old values if possible, we don't care
 		// if here returns an error.
-		rr, err = s.FetchAlias(ctx, r.Alias)
-		if err == nil {
+		//
+		// Note that this is not atomic, meaning that we might run
+		// into concurrent inconsistent issue. But for small scale
+		// use, it is fine for now.
+		rr, err = s.FetchAlias(ctx, a)
+		if err != nil {
+			err = nil
+		} else {
 			// use old values if not presents
 			if r.URL == "" {
 				r.URL = rr.URL
@@ -180,24 +195,29 @@ func shortCmd(ctx context.Context, operate op, r *models.Redirect) (err error) {
 			if rr.Kind != r.Kind {
 				r.Kind = rr.Kind
 			}
+			r.ID = rr.ID
 		}
-		err = nil
+
+		if r.ID == "" {
+			err = fmt.Errorf("cannot find alias %s for update", a)
+			return
+		}
 
 		// do update
 		err = s.UpdateAlias(ctx, r)
 		if err != nil {
 			return
 		}
-		log.Printf("alias %v has been updated.\n", r.Alias)
+		log.Printf("alias %v has been updated.\n", a)
 	case opDelete:
-		err = s.DeleteAlias(ctx, r.Alias)
+		err = s.DeleteAlias(ctx, a)
 		if err != nil {
 			return
 		}
-		log.Printf("alias %v has been deleted.\n", r.Alias)
+		log.Printf("alias %v has been deleted.\n", a)
 	case opFetch:
 		var r *models.Redirect
-		r, err = s.FetchAlias(ctx, r.Alias)
+		r, err = s.FetchAlias(ctx, a)
 		if err != nil {
 			return
 		}
@@ -210,77 +230,168 @@ func shortCmd(ctx context.Context, operate op, r *models.Redirect) (err error) {
 // found in the redir store.
 func (s *server) shortHandler(kind models.AliasKind) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := r.Context()
-
 		var err error
 		defer func() {
 			if err != nil {
-				// Just tell the user we could not find the record rather than
-				// throw 50x. The server should be able to identify the issue.
-				log.Printf("stats err: %v\n", err)
-				// Use 307 redirect to 404 page
+				// Just redirect the user we could not find the record rather than
+				// throw 50x. The server logs should be able to identify the issue.
+				log.Printf("request err: %v\n", err)
 				http.Redirect(w, r, "/404.html", http.StatusTemporaryRedirect)
 			}
 		}()
 
-		// statistic page
-		var prefix string
-		switch kind {
-		case models.KindShort:
-			prefix = conf.S.Prefix
-		case models.KindRandom:
-			prefix = conf.R.Prefix
+		switch r.Method {
+		case http.MethodPost:
+			err = s.shortHandlerPost(kind, w, r)
+		case http.MethodGet:
+			err = s.shortHandlerGet(kind, w, r)
+		default:
+			err = fmt.Errorf("%s is not supported", r.Method)
 		}
+	})
+}
 
-		alias := strings.TrimSuffix(strings.TrimPrefix(r.URL.Path, prefix), "/")
-		if alias == "" {
-			err = s.stats(ctx, kind, w, r)
-			return
-		}
+type shortInput struct {
+	Op    op          `json:"op"`
+	Alias string      `json:"alias"`
+	Data  interface{} `json:"data"`
+}
 
-		// figure out redirect location
-		red, ok := s.cache.Get(alias)
-		if !ok {
-			red, err = s.checkdb(ctx, alias)
-			if err != nil {
-				red, err = s.checkvcs(ctx, alias)
-				if err != nil {
-					return
-				}
-			}
-			s.cache.Put(alias, red)
-		}
+// shortHandlerPost handles all kinds of operations.
+// This is not a RESTful style, because we don't have that much router space
+// to use. We are currently limited the single index router, which is the /s.
+func (s *server) shortHandlerPost(kind models.AliasKind, w http.ResponseWriter, r *http.Request) (err error) {
+	// All post request must be authenticated.
+	err = s.handleAuth(w, r)
+	if err != nil {
+		return
+	}
 
-		// redirect the user immediate, but run pv/uv count in background
-		if time.Now().UTC().Sub(red.ValidFrom.UTC()) > 0 {
-			http.Redirect(w, r, red.URL, http.StatusTemporaryRedirect)
-		} else {
-			err = sTmpl.Execute(w, &struct {
-				ValidFrom time.Time
-			}{red.ValidFrom})
+	// Decode request body and determine what is the operator
+	d := json.NewDecoder(r.Body)
+	var red shortInput
+	err = d.Decode(&red)
+	if err != nil {
+		return
+	}
+
+	// Validating the operator and decode the redir data
+	if !op(red.Op).valid() {
+		err = errors.New("unsupported operator")
+		return
+	}
+
+	b, err := json.Marshal(red.Data)
+	if err != nil {
+		return
+	}
+
+	var redir models.Redirect
+	err = json.Unmarshal(b, &redir)
+	if err != nil {
+		return
+	}
+
+	// Edit redirect data.
+	err = shortEdit(r.Context(), s.db, op(red.Op), red.Alias, &redir)
+	return
+}
+
+func (s *server) shortHandlerGet(kind models.AliasKind, w http.ResponseWriter, r *http.Request) (err error) {
+	ctx := r.Context()
+
+	// statistic page
+	var prefix string
+	switch kind {
+	case models.KindShort:
+		prefix = conf.S.Prefix
+	case models.KindRandom:
+		prefix = conf.R.Prefix
+	}
+
+	alias := strings.TrimSuffix(strings.TrimPrefix(r.URL.Path, prefix), "/")
+
+	// Process visitor information, wait maximum 5 seconds.
+	recordCtx, cancel := context.WithTimeout(ctx, 5*time.Second)
+	defer cancel()
+	s.recognizeVisitor(recordCtx, w, r, alias, kind)
+
+	// If alias is empty, then process index page request.
+	if alias == "" {
+		err = s.sIndex(ctx, kind, w, r)
+		return err
+	}
+
+	// Figure out redirect location
+	red, ok := s.cache.Get(alias)
+	if !ok {
+		red, err = s.checkdb(ctx, alias)
+		if err != nil {
+			red, err = s.checkvcs(ctx, alias)
 			if err != nil {
 				return
 			}
 		}
+		s.cache.Put(alias, red)
+	}
 
-		// count visit in another goroutine so it won't block the redirect.
-		go func() {
-			ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-			defer cancel()
+	// redirect the user immediate, but run pv/uv count in background
+	if time.Now().UTC().Sub(red.ValidFrom.UTC()) > 0 {
+		http.Redirect(w, r, red.URL, http.StatusTemporaryRedirect)
+	} else {
+		err = sTmpl.Execute(w, &struct {
+			ValidFrom time.Time
+		}{red.ValidFrom})
+		if err != nil {
+			return
+		}
+	}
 
-			err := s.db.RecordVisit(ctx, &models.Visit{
-				Alias:   alias,
-				Kind:    kind,
-				IP:      readIP(r),
-				UA:      r.UserAgent(),
-				Referer: r.Referer(),
-				Time:    time.Now().UTC(),
-			})
-			if err != nil {
-				log.Printf("cannot record %s visit: %v", alias, err)
-			}
-		}()
+	return
+}
+
+const redirVidCookie = "redir_vid"
+
+// recognizeVisitor implements a best effort visitor recording.
+//
+// If the redir's cookie is presented, then we use cookie id.
+// If the cookie does not present any data, we read the IP address, and
+// allocates a new visitor id for the visitor.
+//
+// We don't care if any error happens inside.
+func (s *server) recognizeVisitor(
+	ctx context.Context,
+	w http.ResponseWriter,
+	r *http.Request,
+	alias string,
+	kind models.AliasKind,
+) {
+	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
+	defer cancel()
+
+	var cookieVid string
+	c, err := r.Cookie(redirVidCookie)
+	if err != nil {
+		cookieVid = ""
+	} else {
+		cookieVid = c.Value
+	}
+
+	// count visit and set cookie.
+	vid, err := s.db.RecordVisit(ctx, &models.Visit{
+		VisitorID: cookieVid,
+		Alias:     alias,
+		Kind:      kind,
+		IP:        utils.ReadIP(r),
+		UA:        r.UserAgent(),
+		Referer:   r.Referer(),
+		Time:      time.Now().UTC(),
 	})
+	if err != nil {
+		log.Printf("cannot record alias %s's visit: %v", alias, err)
+	} else {
+		w.Header().Set("Set-Cookie", redirVidCookie+"="+vid)
+	}
 }
 
 // checkdb checks whether the given alias is exsited in the redir database
@@ -348,32 +459,41 @@ type records struct {
 	Records []models.VisitRecord
 }
 
-const (
-	username = "x"
-	password = "x"
-)
-
 func (s *server) handleAuth(w http.ResponseWriter, r *http.Request) error {
-	w.Header().Set("WWW-Authenticate", `Basic realm=""`)
+	if !conf.Auth.Enable {
+		return nil
+	}
+
+	w.Header().Set("WWW-Authenticate", `Basic realm="redir"`)
 
 	u, p, ok := r.BasicAuth()
 	if !ok {
-		w.WriteHeader(401)
+		w.WriteHeader(http.StatusUnauthorized)
 		return errors.New("failed to parsing basic auth")
 	}
-	if u != username {
-		w.WriteHeader(401)
+	if u != conf.Auth.Username {
+		w.WriteHeader(http.StatusUnauthorized)
 		return fmt.Errorf("username is invalid: %s", u)
 	}
-	if p != password {
-		w.WriteHeader(401)
+	if p != conf.Auth.Password {
+		w.WriteHeader(http.StatusUnauthorized)
 		return fmt.Errorf("password is invalid: %s", p)
 	}
-	w.WriteHeader(200)
 	return nil
 }
 
-func (s *server) stats(ctx context.Context, kind models.AliasKind, w http.ResponseWriter, r *http.Request) error {
+// sIndex serves two types of index page, and serves statistics data.
+//
+// If there are no supplied value of a `mode` query parameter, the method
+// returns a public visible index page that contains all publicly visible
+// short urls.
+//
+// If the query parameter contains mode=admin, then it requires basic
+// auth to access the admin dashboard where one can manage all short urls.
+//
+// If the query parameter contaisn mode=stat, then it returns application/json
+// data, which contains data for data visualizations in the index page.
+func (s *server) sIndex(ctx context.Context, kind models.AliasKind, w http.ResponseWriter, r *http.Request) error {
 	mode := r.URL.Query().Get("mode")
 	switch mode {
 	case "admin":
@@ -405,7 +525,7 @@ func (s *server) stats(ctx context.Context, kind models.AliasKind, w http.Respon
 		Prefix:  prefix,
 		Records: nil,
 	}
-	rs, err := s.db.CountVisit(ctx, kind)
+	rs, err := s.db.StatVisits(ctx, kind)
 	if err != nil {
 		return err
 	}
@@ -450,19 +570,19 @@ func (s *server) statData(
 	var results interface{}
 	switch stat {
 	case "referer":
-		results, err = s.db.CountReferer(ctx, a, k, start, end)
+		results, err = s.db.StatReferer(ctx, a, k, start, end)
 		if err != nil {
 			retErr = err
 			return
 		}
 	case "ua":
-		results, err = s.db.CountUA(ctx, a, k, start, end)
+		results, err = s.db.StatUA(ctx, a, k, start, end)
 		if err != nil {
 			retErr = err
 			return
 		}
 	case "time":
-		results, err = s.db.CountVisitHist(ctx, a, k, start, end)
+		results, err = s.db.StatVisitHist(ctx, a, k, start, end)
 		if err != nil {
 			retErr = err
 			return
