+++ test.json
@@ -1,0 +1,10 @@
+{
+    "op": "create",
+    "alias": "hello",
+    "data": {
+        "alias": "hello",
+        "url": "https://github.com/changkun",
+        "private": true,
+        "valid_from": "2022-01-01T00:00:00+00:00"
+    }
+}
